update dataset generation with new min feasible sample approach
+ number of network agents is no longer fixed and should not be an arg
+ hdf5 create_dataset requires a fixed shape; however, the number of comm agents
in each sample is unknown and we don't have all the samples in memory at the
same time (to, e.g., compute the size of the largest team). the number of comm
agents comes from min_feasible_config and an estimate of the worst case number
of agents to form a minimum spanning tree is: 3 * ceil(space_side_length /
comm_range). thus, we can safely assume each comm config is <= to this number
and can be padded with NaNs so that every generated sample has a fixed shape for
the hdf5 container