ensure generated training data utilizes the whole image
Since the image size is based on the maximum number of agents expected to be in
the space, the bounding box within which samples are drawn for smaller teams
represents a subset of the entire image. However, we want all areas of the image
to be utilized in the training data. Thus, each configuration is randomly
shifted around the image so that all image extents are visited regardles of team
size.