improve training data
the more sparse the training data the more likely the connectivity optimization
cannot improve on the minimum spanning tree solution, which is an undesired
quality in the training data; by shrinking the sampling area, the node positions
are condensed and the resulting connectivity solutions are more
interesting (i.e. not just on a straight line between task agents given by the
MST solution). NOTE this could be equivalently achieved by increasing the
communication range of the channel model