add coverage range to lloyd's algorithm and tune parameters
the cuttoff range ensures that the centroid of each voronoi cell is only
computed with respect to cells that are nearby the agent; this has the effect of
drawing the robot towards the highest intensity points in its range and prevents
the agent from being skewed away from the peak by distant intensity regions that
fall within the cell. reducing the threshold for peak extraction helps with
coverage (more peaks are detected -> more agents deployed -> more coverate)
however it is not a final solution. some heuristic method should still be
applied that checks if there are large uncovered regions.