removing mid repo as a submodule
The motivation behind adding 'mid' as a submodule was to make this repo entirely
separated from ROS; however, 'mid' contains too many ROS related imports for
this to be possible. Thus, even as a submodule it requires the existence of ROS
related packages on the python path, which necessitates a ROS installation on
the machine. Might as well just build 'mid' in a separate catkin package and
source it before running anything here.