change method by which connectivity is computed from CNN outputs
Instead of simply computing connectivity with the given channel model and
returning the result even if it is not connected, we instead find the minimum
required transmit power to connect the configuration. This is a more faithful
way to compute the CNN results since some of the outputs are arbitrarily close
to being connected and require very little additional power (i.e. increased
communication range) to achieve connectivity.