remove redundant agents from CNN configurations
Often the CNN will deploy additional agents creating cycles in the graph. We
find an approximate solution to the steiner tree on graphs (a connected subgraph
that includes a set of terminal nodes -- i.e. task agents -- and minimizes the
edge length) and remove communication agents that are not a part of it.